requirements.txt
fastmcp>=2.12.0
langchain>=0.3.27
langchain-google-genai>=2.0.10
pandas>=2.2.3
pydantic>=2.11.0
python-dotenv>=1.0.1
